Move beacon API interface to routes
