+++ packages/api/src/beacon/index.ts
@@ -1,27 +1,10 @@
-import {Api as BeaconApi} from "./routes/beacon/index.js";
-import {Api as ConfigApi} from "./routes/config.js";
-import {Api as DebugApi} from "./routes/debug.js";
-import {Api as EventsApi} from "./routes/events.js";
-import {Api as LightclientApi} from "./routes/lightclient.js";
-import {Api as LodestarApi} from "./routes/lodestar.js";
-import {Api as NodeApi} from "./routes/node.js";
-import {Api as ValidatorApi} from "./routes/validator.js";
+import {Api} from "./routes/index.js";
 
 // NOTE: Don't export server here so it's not bundled to all consumers
 
 export * as routes from "./routes/index.js";
 export {getClient} from "./client/index.js";
-
-export type Api = {
-  beacon: BeaconApi;
-  config: ConfigApi;
-  debug: DebugApi;
-  events: EventsApi;
-  lightclient: LightclientApi;
-  lodestar: LodestarApi;
-  node: NodeApi;
-  validator: ValidatorApi;
-};
+export {Api};
 
 // Declare namespaces for CLI options
 export type ApiNamespace = keyof Api;
