+++ packages/api/src/beacon/server/index.ts
@@ -1,5 +1,5 @@
 import {IChainForkConfig} from "@chainsafe/lodestar-config";
-import {Api} from "../../interface.js";
+import {Api} from "../routes/index.js";
 import {ServerInstance, ServerRoute, RouteConfig, registerRoute} from "../../utils/server/index.js";
 
 import * as beacon from "./beacon.js";
